+++ src/main/java/com/example/demo/Entidades/Empleado.java
@@ -36,7 +36,7 @@ public class Empleado extends Persona implements Serializable {
     @OneToMany(mappedBy = "empleado", cascade = CascadeType.ALL, orphanRemoval = true)
     private List<Factura> facturas;
 
-    public Empleado(int cedula, String nombre, String apellido, String fechaNacimiento, String correoElectronico, int telefono, String genero, String direccion, Nomina nomina, String Categoria) {
+    public Empleado(Long cedula, String nombre, String apellido, String fechaNacimiento, String correoElectronico, int telefono, String genero, String direccion, Nomina nomina, String Categoria) {
         super(cedula, nombre, apellido, fechaNacimiento, correoElectronico, telefono);
         this.genero = genero;
         this.direccion = direccion;
